Set up Firebase Firestore for use on the backend.
Why:

Need to store user info to a DB once they register, Firebase Firestore
selected for ease of integration with Flutter.

This change addresses the need by:

Creating a server dir and initiliazing a Firebase project within it and
deploying the rules from the locked mode for the Cloud Firestore.