Create Auth class, test signUp
Why:

To test that the correct auth method is being called on sign up.
Modularise code with separate Auth class and separate concerns with
and abstract class. Keep tests strong.

This change addresses the need by:

Create auth class and an abstract base auth class, wrap firebase auth
method in the auth class, mock it using Mockito and check that it is
being called once upon a user tapping the sign up button.