Test error cases for user sign up
Why:

To check that a user is not created if the registration form has not
been filled in correctly.

This change addresses the need by:

Testing that signUp is not called when the form is empty or when their
are validation errors at the time the user taps the sign up button.