Change HomePage to LandingPage and test route
Why:

Clearer naming for pages and to check that on valid sign up the user is
not shown the registration page.

This change addresses the need by:

Change names throughout app. Test that on valid sign up the user is no
longer shown the registration page and implement the navigation away
from the registration page.