Test and implement logic to check user status
Why:

So that the app can tell if a user is signed in or not and display the
correct screens conditional on user auth status.

This change addresses the need by:

Passinf a callback to the registration page and form and calling it on
successful sign up. Using the response from that callback to
conditionally build either the Landing Page or the Home Page. Wrote
associated tests.