Refactor to use Inherited Widget for auth
Why:

So that I can access Auth state and functions from anywhere in the
widget tree rather than passing them round through constructors.

This change addresses the need by:

Implemented an inherited widget and updated the codebase to use that as
well as updating the widget used for registration form tests.