Refactor to use Inherited Wiget for Auth
Why:

Auth and knowledge of user authentication status is needed throughout
the app.

This change addresses the need by:

Used an inherited widget to share auth methods and information on a
user's authentication status between widgets. Refactored old code and
tests to use the new inherited widget and added a widget test to check
that a user is taken to the Home Page once successfully signed up.