User must enter an email address for login
Why:

When a user wants to sign in with their email address they must enter an
email address, if they do not they will be shown a warning message.

This change addresses the need by:

Test that a warning message is shown if no email address is entered and
implement validator functions as used in the registration form.