Validate password field on login form
Why:

To ensure that a user enters a well formed password when they try to
login.

This change addresses the need by:

Test that a user is shown a warning if they attempt to login without
entering a valid password.  Implement validator function on login form
password field and show new warning message for a poorly formed
password.