Improve test cases
Why:

Make the tests more robust to ensure they are testing only what they
need to be.

This change addresses the need by:

Adding in a test case for when a user enters nothing, updating other
tests so that the user enters either a valid email or password to ensure
that the test case is only testing the invalid entry.