Test valid login
Why:

To check that no validator warnings are shown to the user if they enter
a valid email and password, this does not cover whether the user exists
or not.

This change addresses the need by:

Add test to check that no warnings are shown prior to checking whether
the user has signed up previously or not.