Move validation warning messages to own file
Why:

Modularise the code and have one place to change the message to be used
rather than multiple places.

This change addresses the need by:

Created a new file to hold the validation warning messages, update the
form valdiators to use those messages and update tests to use them as
well rather than duplicating in the test helper file.