Sign up fails if email is already registered
Why:

A user cannot sign up with an email address that has already been
registered with the app.

This change addresses the need by:

Test that a form validation warning is shown to the user if they attempt
to sign up with an email address that is already registered by Firebase
Auth. Implement new logic in the form to catch the firebase error and
rerun the validator functions.