Valid sign in navigates user to Home Page
Why:

A user should be taken to their home page when they sign in
successfully. Also some minor refactors and addition of the firebase
auth sign in error message.

This change addresses the need by:

Test that a user is taken to the home page after a valid sign in.
Implemented by popping the login form route off the navigator.
Added the Firebase Auth sign in error message to use in the logic of the
catch block on sign in.
Refactored the auth class to DRY out code.