Catch Firebase sign in errors
Why:

Display warning messages to the user to notify them the reason for a
failed sign in.
Minor refactor for test helper methods and warning messages to make
naming clearer.

This change addresses the need by:

Test that users see warnings for why sign in failed and implement logic
in the catch block from the sign in attempt.
Rename test helper methods and warning messages.