Add route to login page from registration page
Why:

So that users who are already registered can go to the login page if
they ended up at the sign up page by mistake.

This change addresses the need by:

Test that there is a widget that users can click to go to the login page
and that the routing to the login page is correct.
Add a flat button to the login form and add a navigation function to the
onPressed callback.