Button styling
Why:

To make the sign up and login buttons stretch to fill more of the screen
real estate, easier for people to press on their phone screens.

This change addresses the need by:

In the forms the raised buttons were wrapped in a sized box to
control their width. On the landing page a flexible widget was used to
constrain the width a bit more.