Begin implementation of Google Sign In
Why:

So that users can sign in with Google.

This change addresses the need by:

Updating the Auth class so that it has a signInWithGoogle method.
Testing the success case with Google Sign In, mocking out the external
dependencies that it introduces.
Updating the 'renders' test to expect an extra button for the Google
Sign in option.