Add assets dir and make available
Why:

So that image assets can be used in the app, in this case the Google
logo for the sign in with Google option.

This change addresses the need by:

Created an assets folder, added the Google logo file and added the
assets filed to the pubspec.yaml with the route to the assets.