Rename helper file, catch Google sign in error
Why:

Rename helper file to be more generic / clearer. Catch errors thrown on
Google sign in so that the user can check they have entered the correct
details, sign in another way or try again later.

This change addresses the need by:

Test that errors are caught and a message displayed to the user to
inform them, implement the catch block and display a snack bar.
Rename form helper file.