Google Error Snackbar duration and test
Why:

The message displayed to the user in the snack bar should be removed
after a given duration.

This change addresses the need by:

Test that after a given duration the snackbar is removed and is not
visible to the user.  Add a duration widget to the Snackbar.
Add a constant variable to the helper file for that duration.