Add drop down menu to Home Page
Why:

So that users can select from a range of actions such as sign out or
navigation to other parts of the app.

This change addresses the need by:

Testing that another Icon Button is present and implementing the
PopupMenuButton on the Home Page as well as creating a Choice class that
is used to create instances of the choices available on the drop down
menu.