Add user to Firestore on Google Sign in
Why:

A user should be added to Firestore if they sign in to the app with
Google for the first time and the email address used is not already
registered.

This change addresses the need by:

Test that a user document is created in Firestore when Google sign in is
used to sign in to the app.
Implement the call to setData on the users collection in Firestore.
Update the Google sign in wrapper function in the auth services so that
it returns a Map which contains both the uid and the email address for
that user.
Udpate the FirebaseUserMock to have an email property for use in tests.