Do not set user doc if user already exists
Why:

If a user already exists then a write should not be made to the
Firestore document.

This change addresses the need by:

Test that when a user exists that setData is not called on the Firestore
instance. Set up a new mock for DocumentSnapshot. Implement check of
Firestore document before signing in.